requirements.txt
# Core Dependencies


# LLM & AI
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-community>=0.0.20
transformers>=4.36.0
torch>=2.1.0

# RAG & Vector Storage
chromadb>=0.4.22
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1

# Text Processing
nltk>=3.8.1
spacy>=3.7.2
textblob>=0.17.1
unidecode>=1.3.7

# Data & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API & Web
streamlit>=1.31.0
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.5.0

# Configuration & Utils
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Logging & Monitoring
loguru>=0.7.0

# Optional: For additional LLM support
llama-cpp-python>=0.2.0  # For local LLaMA models
mistralai>=0.0.11  # For Mistral API